- (keytools.c, keytools.h, parse.c, scapi.c): Patch from Markku Laukkanen:   - size_t changes.
git-svn-id: file:///home/hardaker/lib/sf-bkups/net-snmp-convert-svnrepo/trunk/net-snmp@2885 06827809-a52a-0410-b366-d66718629ded

+++ snmplib/scapi.c
@@ -117,7 +117,7 @@ sc_get_transform_type(oid *hashtype, u_int hashtype_len,
                       int (**hash_fn)(
                         const int	  mode,		void  	 **context,
                         const u_int8_t	 *data,		const int  data_len,
-                        u_int8_t	**digest,	int	  *digest_len))
+                        u_int8_t	**digest,	size_t  *digest_len))
 {
   /*
    * Determine transform type.
